Cut/paste dataframe tests to integration
